refactor(schema): move collection and user schemas to `@repo/schema`
Note: `collectionCardImageSchema` and `createClientCollectionValidation` were not moved because they have local dependencies.
