+++ apps/next-app/src/components/user/user-info.tsx
@@ -2,7 +2,7 @@
 
 // types
 import type { LucideIcon, LucideProps } from "lucide-react"
-import type { ClientUser } from "@/lib/schema/user-schema"
+import type { ClientUser } from "@repo/schema/user"
 
 // utils
 import { cn } from "@/lib/util"
