+++ apps/next-app/src/lib/schema/collection/index.ts
@@ -1,0 +1,1 @@
+export * from "./collection-schema"
