+++ apps/next-app/src/lib/util/parser/user-parser.ts
@@ -1,5 +1,5 @@
 // types
-import type { ClientUser } from "@/lib/schema/user-schema"
+import type { ClientUser } from "@repo/schema/user"
 
 /* Schema parser keys */
 export const clientUserKeys: (keyof ClientUser)[] = [
