+++ packages/schema/package.json
@@ -10,12 +10,15 @@
     "check-types": "tsc --noEmit"
   },
   "exports": {
+    "./collection": "./dist/collection/index.js",
+    "./collection-validation": "./dist/collection/validation/index.js",
     "./player": "./dist/player/index.js",
     "./player-validation": "./dist/player/validation/index.js",
     "./room": "./dist/room/index.js",
     "./room-validation": "./dist/room/validation/index.js",
     "./session": "./dist/session/index.js",
-    "./session-validation": "./dist/session/validation/index.js"
+    "./session-validation": "./dist/session/validation/index.js",
+    "./user": "./dist/user/index.js"
   },
   "dependencies": {
     "@repo/db": "workspace:*"
