+++ packages/schema/src/collection/collection-schema.ts
@@ -1,0 +1,37 @@
+import { z } from "zod"
+
+// schemas
+import { TableSizeSchema } from "@repo/db/zod"
+import { clientUserSchema } from "@/user"
+
+/* Base schemas */
+export const collectionNameSchema = z.string()
+  .min(4, { message: "Collection name is too short." })
+  .max(28, { message: "Collection name is too long." })
+
+export const collectionDescriptionSchema = z.string()
+  .min(8, { message: "Collection description is too short." })
+  .max(128, { message: "Collection description is too long." })
+
+/* Client base schemas */
+export const clientMemoryCardSchema = z.object({
+  id: z.string(),
+  imageUrl: z.string()
+})
+
+export const clientCollectionSchema = z.object({
+  id: z.string(),
+  name: collectionNameSchema,
+  description: collectionDescriptionSchema,
+  tableSize: TableSizeSchema,
+  user: clientUserSchema,
+  cards: z.array(clientMemoryCardSchema),
+
+  createdAt: z.coerce.date(),
+  updatedAt: z.coerce.date()
+})
+
+export type CollectionName = z.infer<typeof collectionNameSchema>
+export type CollectionDescription = z.infer<typeof collectionDescriptionSchema>
+export type ClientMemoryCard = z.infer<typeof clientMemoryCardSchema>
+export type ClientCardCollection = z.infer<typeof clientCollectionSchema>
