+++ packages/schema/src/player/validation/player-action-validation.ts
@@ -1,7 +1,7 @@
 import { z } from "zod"
 
 // schemas
-import { playerColor, playerTag } from "@repo/schema/player"
+import { playerColor, playerTag } from "@/player"
 
 export const createPlayerValidation = z.object({
   tag: playerTag,
